move interpreter code to a separate file
